data-services/requirements.txt
pandas
scikit-learn
psycopg2-binary
sqlalchemy
pyyaml
tabula-py
boto3
numpy